res fixed a few VC7 64bit portatibility warnings and put down comments   like "@@@ FIXME 64bit" at problematic to-be-fixed places. res enabled VC7 64bit portatibility warnings for libraries.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18346 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/bitarray.h
@@ -109,7 +109,7 @@ public:
     size_t mPos;
   public:
     /// Constructor.
-    BitProxy(csBitArray &array, unsigned pos): mArray(array), mPos(pos)
+    BitProxy(csBitArray &array, size_t pos): mArray(array), mPos(pos)
     {}
 
     /// Boolean assignment.
