+++ plugins/video/loader/jpg/jpgimage.h
@@ -40,7 +40,7 @@ class csJPGImageIO : public iImageIO
   virtual ~csJPGImageIO ();
 
   virtual const csImageIOFileFormatDescriptions& GetDescription ();
-  virtual csPtr<iImage> Load (uint8* iBuffer, uint32 iSize, int iFormat);
+  virtual csPtr<iImage> Load (uint8* iBuffer, size_t iSize, int iFormat);
   virtual void SetDithering (bool iEnable);
   virtual csPtr<iDataBuffer> Save (iImage *image, const char *mime = 0,
     const char* extraoptions = 0);
@@ -74,7 +74,7 @@ private:
   ImageJpgFile (int iFormat, iObjectRegistry* p) : csImageFile (iFormat) 
     { object_reg = p; };
   /// Try to read the PNG file from the buffer and return success status
-  bool Load (uint8* iBuffer, uint32 iSize);
+  bool Load (uint8* iBuffer, size_t iSize);
 };
 
 #endif // __CS_JPGIMAGE_H__
