+++ plugins/video/loader/wal/walimage.h
@@ -40,7 +40,7 @@ class csWALImageIO : public iImageIO
   virtual ~csWALImageIO ();
 
   virtual const csImageIOFileFormatDescriptions& GetDescription ();
-  virtual csPtr<iImage> Load (uint8* iBuffer, uint32 iSize, int iFormat);
+  virtual csPtr<iImage> Load (uint8* iBuffer, size_t iSize, int iFormat);
   virtual void SetDithering (bool iEnable);
   virtual csPtr<iDataBuffer> Save (iImage *image, const char *mime = 0,
     const char* extraoptions = 0);
@@ -64,7 +64,7 @@ private:
   /// Initialize the image object
   ImageWALFile (int iFormat) : csImageFile (iFormat) { };
   /// Try to read the WAL file from the buffer and return success status
-  bool Load (uint8* iBuffer, uint32 iSize);
+  bool Load (uint8* iBuffer, size_t iSize);
 };
 
 #endif // __CS_WALIMAGE_H__
